+++ pubspec.yaml
@@ -43,6 +43,7 @@ dependencies:
 
 
 
+
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
@@ -75,11 +76,13 @@ flutter:
   #   - images/a_dot_ham.jpeg
   assets:
     - assets/logo.png
+    - assets/audio/alarm.mp3
     - assets/quiz_images/question 1.png
     - assets/quiz_images/question 2.png
     - assets/quiz_images/question 3.png
     - assets/quiz_images/question 4.png
     - assets/quiz_images/question 5.png
+    - assets/audio/Chicago - Freedom Trail Studio.mp3
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware
